Configure YUM proxy on instance launch
YUM will need to have the proxy set to the correct value based on which env
the instance is being launched in (e.g. management-dev proxy for dev instances,
management proxy for production instances).

We lay down a simple(ish) shell script that retrieves the proxy endpoint
specified in the `internet_proxy` instance tag, then configure the proxy for
each individual YUM repo that needs it.

Note that we explicitly *don't* set a global YUM default proxy in /etc/yum.conf
because repositories that are laid down on the instances themselves (e.g. EMR)
could well be S3 based, and therefore *mustn't* use the proxy.