Enable YUM proxy setting to run under EMR (#178)
* Enable YUM proxy setting to run under EMR

EMR specifies a `runcmd` section in the userdata of the instances that it
spins up. That overrides any `runcmd` sections we have on disk. Instead,
specify that the yum proxy setting script be run during `bootcmd` instead.

This is much earlier than we'd generally want things like this to be run,
but the default AL1 AMI doesn't specify any `bootcmd` sections, neither does
EMR's userdata so we avoid any conflicts/overrides by doing things this way.

* Fix cloud-init-per syntax